# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teleasistencia Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/incidencias": {
            "get": {
                "tags": [
                    "Incidencias"
                ],
                "summary": "Obtener todas las incidencias",
                "operationId": "d7e2a4aeb9d8709d7add7dce206f1369",
                "responses": {
                    "200": {
                        "description": "Lista de incidencias recuperada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/IncidentResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Incidencias"
                ],
                "summary": "Crear una nueva incidencia",
                "operationId": "9d35933491df7b751657e0a14de37134",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreIncidentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Incidencia creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IncidentResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/incidencias/{id}": {
            "get": {
                "tags": [
                    "Incidencias"
                ],
                "summary": "Obtener detalles de una incidencia",
                "operationId": "bb0cdc60566dae1ee845db7ba5818e04",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la incidencia recuperados exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IncidentResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Incidencias"
                ],
                "summary": "Actualizar una incidencia",
                "operationId": "eaf9a1ad1d537c255ba0430e7ac977d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateIncidentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Incidencia actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IncidentResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Incidencias"
                ],
                "summary": "Eliminar una incidencia",
                "operationId": "acb7799fd741a258adefacee3184bb7d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Incidencia eliminada exitosamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/incidencias/paciente/{patient}": {
            "get": {
                "tags": [
                    "Incidencias"
                ],
                "summary": "Obtener todas las incidencias de un paciente",
                "operationId": "812dcd8624297a8655ceb7318d033127",
                "parameters": [
                    {
                        "name": "patient_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de incidencias del paciente recuperada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/IncidentResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreIncidentRequest": {
                "description": "Validación para la creación de incidencias",
                "required": [
                    "paciente_id"
                ],
                "properties": {
                    "paciente_id": {
                        "description": "ID del paciente asociado a la incidencia",
                        "type": "integer",
                        "example": 5
                    },
                    "descripcion": {
                        "description": "Descripción detallada de la incidencia",
                        "type": "string",
                        "example": "El paciente sufrió una caída durante la noche",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UpdateIncidentRequest": {
                "description": "Validación para la actualización de incidencias",
                "properties": {
                    "paciente_id": {
                        "description": "ID del paciente asociado a la incidencia",
                        "type": "integer",
                        "example": 5
                    },
                    "descripcion": {
                        "description": "Descripción detallada de la incidencia",
                        "type": "string",
                        "example": "El paciente experimentó una caída durante la noche"
                    }
                },
                "type": "object"
            },
...
            "IncidentResource": {
                "description": "Esquema del recurso Incidencia",
                "properties": {
                    "id": {
                        "description": "Identificador de la incidencia",
                        "type": "integer",
                        "example": 1
                    },
                    "paciente_id": {
                        "description": "ID del paciente asociado a la incidencia",
                        "type": "integer",
                        "example": 5
                    },
                    "descripcion": {
                        "description": "Descripción detallada de la incidencia",
                        "type": "string",
                        "example": "El paciente presentó una caída durante la noche",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Fecha de creación de la incidencia",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-23 12:00:00"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización de la incidencia",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-24 12:00:00"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}